Brings back jsx processing
Not every use case of using JSX and React with Meteor warrants using the
`reactjs:react` package. That package encompasses a lot of concerns that
can be uncoupled -- loading React, the JSX compiler, and some
React mixins. Those can be decoupled to provide better modularity,
of which processing JSX files with this package is one example